Transform or convert coordinates of simple features directly with Proj.4

Transform or convert coordinates of simple features directly with Proj.4

st_transform_proj(x, crs, ...)

# S3 method for sfc
st_transform_proj(x, crs, ...)

# S3 method for sf
st_transform_proj(x, crs, ...)

# S3 method for sfg
st_transform_proj(x, crs, ...)

Arguments

x
object of class sf, sfc or sfg

crs
character; proj4string of target coordinate reference system

...
ignored

Details

Transforms coordinates of object to new projection, using Proj.4 and not the GDAL API.

The st_transform_proj method for sfg objects assumes that the CRS of the object is available as an attribute of that name.

Examples

library(sf)
p1 = st_point(c(7,52))
p2 = st_point(c(-30,20))
sfc = st_sfc(p1, p2, crs = 4326)
sfc

#> Geometry set for 2 features 
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -30 ymin: 20 xmax: 7 ymax: 52
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs

#> POINT (7 52)

#> POINT (-30 20)

st_transform_proj(sfc, "+proj=wintri")

#> Geometry set for 2 features 
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -2663827 ymin: 2238779 xmax: 524263.4 ymax: 5789966
#> epsg (SRID):    NA
#> proj4string:    +proj=wintri

#> POINT (524263.423302337 5789966.3100998)

#> POINT (-2663827.47315967 2238779.07725309)

library(sf)
nc = st_read(system.file("shape/nc.shp", package="sf"))

#> Reading layer `nc' from data source `/home/edzer/R/x86_64-pc-linux-gnu-library/3.4/sf/shape/nc.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> epsg (SRID):    4267
#> proj4string:    +proj=longlat +datum=NAD27 +no_defs

st_transform_proj(nc[1,], "+proj=wintri +over")

#> Simple feature collection with 1 feature and 14 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -6776742 ymin: 4191789 xmax: -6736178 ymax: 4233236
#> epsg (SRID):    NA
#> proj4string:    +proj=wintri +over
#>    AREA PERIMETER CNTY_ CNTY_ID NAME  FIPS FIPSNO CRESS_ID BIR74 SID74 NWBIR74
#> 1 0.114     1.442  1825    1825 Ashe 37009  37009        5  1091     1      10
#>   BIR79 SID79 NWBIR79                       geometry
#> 1  1364     0      19 MULTIPOLYGON (((-6760647.75...

st_transform_proj(structure(p1, proj4string = "+init=epsg:4326"), "+init=epsg:3857")

#> POINT (779236.435552915 6800125.45439731)